fix(suite-native): add missing button
