+++ suite-native/module-home/src/components/DashboardNavigationButtons.tsx
@@ -2,7 +2,6 @@ import { useNavigation } from '@react-navigation/native';
 
 import { Box, Button, Divider, VStack } from '@suite-native/atoms';
 import {
-    AccountsImportStackRoutes,
     RootStackParamList,
     RootStackRoutes,
     StackNavigationProps,
@@ -11,28 +10,12 @@ import {
 export const DashboardNavigationButtons = () => {
     const navigation = useNavigation<StackNavigationProps<RootStackParamList, RootStackRoutes>>();
 
-    const handleImportAssets = () => {
-        navigation.navigate(RootStackRoutes.AccountsImport, {
-            screen: AccountsImportStackRoutes.SelectNetwork,
-        });
-    };
-
     const handleReceive = () => {
         navigation.navigate(RootStackRoutes.ReceiveModal, {});
     };
 
     return (
         <VStack spacing="large">
-            <Box marginHorizontal="medium">
-                <Button
-                    data-testID="@home/portfolio/sync-coins-button"
-                    colorScheme="tertiaryElevation0"
-                    size="large"
-                    onPress={handleImportAssets}
-                >
-                    Sync my coins
-                </Button>
-            </Box>
             <Divider />
             <Box marginHorizontal="medium">
                 <Button
